[GFS2] Cast 64 bit printk args to unsigned long long.

+++ fs/gfs2/locking/dlm/plock.c
@@ -94,7 +94,8 @@ int gdlm_plock(lm_lockspace_t *lockspace, struct lm_lockname *name,
 	if (!rv) {
 		if (posix_lock_file_wait(file, fl) < 0)
 			log_error("gdlm_plock: vfs lock error %x,%llx",
-				  name->ln_type, name->ln_number);
+				  name->ln_type,
+				  (unsigned long long)name->ln_number);
 	}
 
 	kfree(op);
@@ -114,7 +115,7 @@ int gdlm_punlock(lm_lockspace_t *lockspace, struct lm_lockname *name,
 
 	if (posix_lock_file_wait(file, fl) < 0)
 		log_error("gdlm_punlock: vfs unlock error %x,%llx",
-			  name->ln_type, name->ln_number);
+			  name->ln_type, (unsigned long long)name->ln_number);
 
 	op->info.optype		= GDLM_PLOCK_OP_UNLOCK;
 	op->info.pid		= fl->fl_pid;
@@ -244,7 +245,7 @@ static ssize_t dev_write(struct file *file, const char __user *u, size_t count,
 		wake_up(&recv_wq);
 	else
 		printk(KERN_INFO "gdlm dev_write no op %x %llx\n", info.fsid,
-			info.number);
+			(unsigned long long)info.number);
 	return count;
 }
